[CL-1307] Split up constants into typings and styling
